shfmt (#21)
Ran `shfmt -w btrfs-auto-snapshot`. I am a fan of having a formatter
for the source files. Let me know what you think, I obviously respect if
you don't care as much (and we leave it as-is).

Also, as I am not a huge fan of third-party github actions, I opted to
install `shellcheck` to the ubuntu image and run it from there.

Diff best previewed in `git diff -w --word-diff`.